Move WPF Windows to a subfolder for grouping.  Add a global error handler.
