+++ KitchenTimer/Windows/MainWindow.xaml
@@ -1,6 +1,6 @@
-﻿<Window x:Class="KitchenTimer.MainWindow"
+﻿<Window x:Class="KitchenTimer.Windows.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        Icon="icons/Timer_16x.png"
+        Icon="../icons/Timer_16x.png"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
