+++ KitchenTimer/Windows/SettingsWindow.xaml
@@ -1,11 +1,11 @@
-﻿<Window x:Class="KitchenTimer.SettingsWindow"
+﻿<Window x:Class="KitchenTimer.Windows.SettingsWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:KitchenTimer"
         xmlns:xctk="clr-namespace:Xceed.Wpf.Toolkit;assembly=Xceed.Wpf.Toolkit"
-        Icon="icons/Timer_16x.png"
+        Icon="../icons/Timer_16x.png"
         WindowStyle="ToolWindow"
         mc:Ignorable="d"
         Title="Timer Settings" Height="139" Width="380" Background="#FF171197">
